refactor runtime health helpers
